requirements.txt
discord.py == 2.2.2
gspread
oauth2client
black